+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/ImageMode/ImageModelCamera.ts
